Rename child with timer
