+++ src/timer.jsx
@@ -3,14 +3,14 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 
-import { configureStore } from './stores/child'
-import ChildApp from './components/timerwindow/ChildApp'
+import { configureStore } from './stores/timer'
+import TimerApp from './components/timerwindow/TimerApp'
 
 const store = configureStore()
 
 ReactDOM.render(
   <Provider store={ store }>
-    <ChildApp />
+    <TimerApp />
   </Provider>,
-  document.getElementById('slippr-child')
+  document.getElementById('slippr-timer')
 )
