+++ timer.html
@@ -8,8 +8,8 @@
     <title>Slippr</title>
   </head>
   <body>
-    <div id="slippr-child"></div>
+    <div id="slippr-timer"></div>
   </body>
 
-  <script src="./dist/child.bundle.js"></script>
+  <script src="./dist/timer.bundle.js"></script>
 </html>
